Move additional helpers into span_ops header
