+++ include/beman/big_int/detail/span_ops.hpp
@@ -5,6 +5,9 @@
 #define BEMAN_BIG_INT_SPAN_OPS_HPP
 
 #include <beman/big_int/detail/config.hpp>
+#include <algorithm>
+#include <ranges>
+#include <utility>
 #include <span>
 
 namespace beman::big_int::detail {
@@ -212,6 +215,164 @@ constexpr std::size_t multiply_single_limb(const std::span<uint_multiprecision_t
 
     return a.size();
 }
+
+// ---------------------------------------------------------------------------
+// Helpers shared by the Toom-Cook variants. All operate on little-endian
+// multi-precision spans of uint_multiprecision_t limbs.
+// ---------------------------------------------------------------------------
+
+// In-place tmp[0..size) += addend; returns new size (may grow by 1).
+[[nodiscard]] constexpr std::size_t add_into_tmp(const std::span<uint_multiprecision_t>       tmp,
+                                                 const std::size_t                            size,
+                                                 const std::span<const uint_multiprecision_t> addend) noexcept {
+    if (addend.empty()) {
+        return size;
+    }
+
+    const std::size_t new_size = std::max(size, addend.size());
+    BEMAN_BIG_INT_DEBUG_ASSERT(tmp.size() > new_size);
+
+    const bool carry =
+        add_unsigned_spans(tmp.first(new_size), std::span<const uint_multiprecision_t>{tmp.data(), size}, addend);
+
+    if (carry) {
+        tmp[new_size] = 1;
+        return new_size + 1;
+    }
+
+    return new_size;
+}
+
+// In-place tmp[0..size) <<= n (0 < n < limb_width); returns new size (may grow by 1).
+// Pass size == tmp.size() to shift the full span; the carry-branch assertion
+// then enforces that the shift did not overflow.
+// Single-pass equivalent of n chained shift_left_one calls.
+[[nodiscard]] constexpr std::size_t
+shift_left_n(const std::span<uint_multiprecision_t> tmp, std::size_t size, const unsigned n) noexcept {
+    if (size == 0 || n == 0) {
+        return size;
+    }
+
+    constexpr std::size_t local_limb_bits = width_v<uint_multiprecision_t>;
+    BEMAN_BIG_INT_DEBUG_ASSERT(size <= tmp.size());
+    BEMAN_BIG_INT_DEBUG_ASSERT(n < local_limb_bits);
+
+    uint_multiprecision_t prev = 0;
+    for (std::size_t i = 0; i < size; ++i) {
+        const auto limb = tmp[i];
+        tmp[i]          = funnel_shl(wide<uint_multiprecision_t>{.low_bits = prev, .high_bits = limb}, n);
+        prev            = limb;
+    }
+
+    if (const auto carry = prev >> (local_limb_bits - n)) {
+        BEMAN_BIG_INT_DEBUG_ASSERT(size < tmp.size());
+        tmp[size++] = carry;
+    }
+
+    return size;
+}
+
+// In-place tmp[0..size) <<= 1; returns new size (may grow by 1). Thin wrapper
+// around shift_left_n for the common single-bit Horner-style shift.
+[[nodiscard]] constexpr std::size_t shift_left_one(const std::span<uint_multiprecision_t> tmp,
+                                                   const std::size_t                      size) noexcept {
+    return shift_left_n(tmp, size, 1u);
+}
+
+// In-place tmp >>= n (0 < n < limb_width); returns the dropped low n bits packed
+// as a single value (caller asserts == 0 for exact division by 2^n).
+// Single-pass equivalent of n chained shift_right_one calls.
+[[nodiscard]] constexpr uint_multiprecision_t shift_right_n(const std::span<uint_multiprecision_t> tmp,
+                                                            const unsigned                         n) noexcept {
+    if (tmp.empty() || n == 0) {
+        return 0;
+    }
+
+    constexpr std::size_t local_limb_bits = width_v<uint_multiprecision_t>;
+    BEMAN_BIG_INT_DEBUG_ASSERT(n < local_limb_bits);
+
+    const uint_multiprecision_t mask = (uint_multiprecision_t{1} << n) - uint_multiprecision_t{1};
+    const uint_multiprecision_t rem  = tmp[0] & mask;
+    uint_multiprecision_t       high = 0;
+    for (std::size_t i = tmp.size(); i-- > 0;) {
+        const auto limb = tmp[i];
+        tmp[i]          = funnel_shr(wide<uint_multiprecision_t>{.low_bits = limb, .high_bits = high}, n);
+        high            = limb;
+    }
+
+    return rem;
+}
+
+// In-place tmp >>= 1; returns the dropped low bit (caller asserts == 0 for exact div).
+// Thin wrapper around shift_right_n for the common single-bit halving case.
+[[nodiscard]] constexpr uint_multiprecision_t shift_right_one(const std::span<uint_multiprecision_t> tmp) noexcept {
+    return shift_right_n(tmp, 1u);
+}
+
+// In-place result[shift..) += src; asserts no carry out of the result span.
+constexpr void add_shifted(const std::span<uint_multiprecision_t>       result,
+                           const std::size_t                            shift,
+                           const std::span<const uint_multiprecision_t> src) noexcept {
+    const auto dest  = result.subspan(shift);
+    bool       carry = false;
+    for (std::size_t i = 0; i < dest.size(); ++i) {
+        const auto si            = i < src.size() ? src[i] : uint_multiprecision_t{0};
+        const auto [r_value, c1] = carrying_add(dest[i], si, carry);
+        dest[i]                  = r_value;
+        carry                    = c1;
+    }
+
+    BEMAN_BIG_INT_DEBUG_ASSERT(!carry);
+}
+
+// Given s = a + b and (|a - b|, sign_d) for non-negative a and b, recover the
+// pair (a, b) and place them in (lower_dst, higher_dst):
+//   sign_d == false  =>  a >= b: lower_dst = a = (s + |d|)/2; higher_dst = b = (s - |d|)/2
+//   sign_d == true   =>  a <  b: lower_dst = a = (s - |d|)/2; higher_dst = b = (s + |d|)/2
+// Used by Toom-Cook palindromic interpolation to split sum/diff back into
+// individual coefficients. `tmp_scratch` is a workspace of at least
+// max(s_view.size(), d_view.size()) limbs; it does not need to be pre-zeroed.
+// The destinations may alias s_view and d_view (recover_pair reads s and d
+// before writing the destinations); destinations must not alias each other
+// nor tmp_scratch. Both s_view + d_view and s_view - d_view must be even
+// (caller's responsibility, asserted via shift_right_one's rem == 0 check).
+constexpr void recover_pair(const std::span<uint_multiprecision_t>       lower_dst,
+                            const std::span<uint_multiprecision_t>       higher_dst,
+                            const std::span<const uint_multiprecision_t> s_view,
+                            const std::span<const uint_multiprecision_t> d_view,
+                            const bool                                   sign_d,
+                            const std::span<uint_multiprecision_t>       tmp_scratch) noexcept {
+    // Stage (s + |d|) in tmp_scratch first; the subtract-into-destination step
+    // below would otherwise clobber s_view / d_view if either aliases a destination.
+    std::ranges::fill(tmp_scratch, uint_multiprecision_t{0});
+    const bool sum_carry = add_unsigned_spans(tmp_scratch, s_view, d_view);
+    BEMAN_BIG_INT_DEBUG_ASSERT(!sum_carry);
+    const std::size_t plus_size = std::max(s_view.size(), d_view.size());
+    const auto        plus_span = std::span<const uint_multiprecision_t>{tmp_scratch.data(), plus_size};
+
+    // (s - |d|)/2 into the destination chosen by sign.
+    const std::span<uint_multiprecision_t> minus_dst = sign_d ? lower_dst : higher_dst;
+    subtract_unsigned_spans(minus_dst, s_view, d_view);
+    {
+        const auto rem = shift_right_one(minus_dst);
+        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
+    }
+
+    // (s + |d|)/2 into the other destination.
+    const std::span<uint_multiprecision_t> plus_dst = sign_d ? higher_dst : lower_dst;
+    std::ranges::fill(plus_dst, uint_multiprecision_t{0});
+    std::ranges::copy(plus_span, plus_dst.begin());
+    {
+        const auto rem = shift_right_one(plus_dst);
+        BEMAN_BIG_INT_DEBUG_ASSERT(rem == 0);
+    }
+}
+
+// Used by the solve_subsystem lambda in toom_cook_6_5
+struct subsystem_signs {
+    bool outer;
+    bool inner;
+};
 
 } // namespace beman::big_int::detail
